projectCServer/requirements.txt
absl-py==1.3.0
aniso8601==9.0.1
asttokens==2.1.0
astunparse==1.6.3
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
filterpy==1.4.5
Flask==2.2.2
flask-restx==1.0.3
flatbuffers==22.12.6
fonttools==4.38.0
gast==0.4.0
gevent==22.10.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.1
grpcio==1.34.1
h5py==3.1.0
idna==3.4
imageio==2.22.4
importlib-metadata==5.1.0
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.10.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lab==7.2
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
networkx==2.8.8
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
protobuf==3.19.6
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
PyWavelets==1.4.1
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.9.3
seaborn==0.12.1
simplejson==3.18.0
six==1.15.0
smmap==5.0.0
SQLAlchemy==1.4.44
stack-data==0.6.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.1
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.28.0
termcolor==1.1.0
thop==0.1.1.post2209072238
tifffile==2022.10.10
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.64.1
traitlets==5.5.0
txt2tags==3.7
typing-extensions==3.7.4.3
urllib3==1.26.12
waitress==2.1.2
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.12.1
zipp==3.11.0
zope.event==4.6
zope.interface==5.5.2
